# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /caresync/access-requests/{id}/decision:
    post:
      operationId: decideAccessRequest
      tags: [caresync]
      summary: Allow or deny a consent request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessDecision"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
  /caresync/appointments:
    get:
      operationId: listCareSyncAppointments
      tags: [caresync]
      summary: List appointments
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createCareSyncAppointment
      tags: [caresync]
      summary: Book a demo appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    AccessRequest:
      type: object
      required: [id, doctorName, specialization, organization, purpose, requestedAt, status, dataScopes]
      properties:
        id: { type: number }
        doctorName: { type: string }
        specialization: { type: string }
        organization: { type: string }
        purpose: { type: string }
        requestedAt: { type: string }
        status: { type: string }
        dataScopes:
          type: array
          items: { type: string }
    AccessDecision:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [ALLOW, DENY]
    Appointment:
      type: object
      required: [id, doctorName, specialization, date, time, mode, status, fee]
      properties:
        id: { type: number }
        doctorName: { type: string }
        specialization: { type: string }
        date: { type: string }
        time: { type: string }
        mode: { type: string }
        status: { type: string }
        fee: { type: number }
    AppointmentInput:
      type: object
      required: [doctorId, date, time, mode]
      properties:
        doctorId: { type: number }
        date: { type: string }
        time: { type: string }
        mode: { type: string }
...